tutorial 04: Merge solution into notebook
