+++ testsuite/scripts/tutorials/CMakeLists.txt
@@ -37,7 +37,6 @@ tutorial_test(FILE test_04-lattice_boltzmann_part2.py LABELS "gpu")
 tutorial_test(FILE test_04-lattice_boltzmann_part3.py LABELS "gpu")
 tutorial_test(FILE test_04-lattice_boltzmann_part3_solution.py LABELS "gpu")
 tutorial_test(FILE test_04-lattice_boltzmann_part4.py LABELS "gpu")
-tutorial_test(FILE test_04-lattice_boltzmann_part4_solution.py LABELS "gpu")
 tutorial_test(FILE test_05-raspberry_electrophoresis.py LABELS "gpu")
 tutorial_test(FILE test_06-active_matter__flow_field.py LABELS "gpu")
 tutorial_test(FILE test_06-active_matter__rectification_geometry.py)
