+++ testsuite/scripts/tutorials/test_04-lattice_boltzmann_part4.py
@@ -17,6 +17,7 @@
 
 import unittest as ut
 import importlib_wrapper
+import numpy as np
 
 
 tutorial, skipIfMissingFeatures = importlib_wrapper.configure_and_import(
@@ -27,6 +28,12 @@ tutorial, skipIfMissingFeatures = importlib_wrapper.configure_and_import(
 @skipIfMissingFeatures
 class Tutorial(ut.TestCase):
     system = tutorial.system
+
+    def test_flow_profile(self):
+        analytical = tutorial.y_values
+        simulation = tutorial.fluid_velocities[:, 1]
+        rmsd = np.sqrt(np.mean(np.square(analytical - simulation)))
+        self.assertLess(rmsd, 2e-5 * tutorial.AGRID / tutorial.lbf.tau)
 
 
 if __name__ == "__main__":
